requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==4.2.10
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-timezone-field==6.1.0
djangorestframework==3.14.0
drf-yasg==1.21.7
idna==3.6
inflection==0.5.1
kombu==5.3.5
packaging==23.2
prompt-toolkit==3.0.43
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
